+++ src/commands/trainer/trainerInfo.js
@@ -1,32 +1,37 @@
-const { getTrainer } = require('../../services/trainer');
+const { getTrainerInfo } = require('../../services/trainer');
 const { buildTrainerEmbed } = require('../../embeds/trainerEmbeds');
 
 const trainerInfo = async (user) => {
-    const trainer = await getTrainer(user);
+    const trainer = await getTrainerInfo(user);
     if (trainer.err) {
         return { embed: null, err: trainer.err };
     }
+
     const embed = buildTrainerEmbed(trainer.data);
-    return { embed: embed, err: null };
+
+    const send = {
+        embeds: [embed]
+    }
+    return { send: send, err: null };
 }
 
 const trainerInfoMessageCommand = async (message) => {
-    const { embed, err } = await trainerInfo(message.author);
+    const { send, err } = await trainerInfo(message.author);
     if (err) {
         await message.channel.send(`${err}`);
         return { err: err };
     } else {
-        await message.channel.send({ embeds: [embed] });
+        await message.channel.send(send);
     }
 }
 
 const trainerInfoSlashCommand = async (interaction) => {
-    const { embed, err } = await trainerInfo(interaction.user);
+    const { send, err } = await trainerInfo(interaction.user);
     if (err) {
         await interaction.reply(`${err}`);
         return { err: err };
     } else {
-        await interaction.reply({ embeds: [embed] });
+        await interaction.reply(send);
     }
 }